Add alt property to logo
