--- src/media/dummyData.js
@@ -1,100 +1,0 @@
-import { makeStyles } from "@material-ui/styles"
-
-export const needsDummy = [
-    {
-        urgencyLevel: 1,
-        itemName: "N95 makeStyles",
-        quantity: "50,000",
-        budget: "Can buy",
-        specifications: "New York City",
-        createdAt: "NY hospital"
-    },
-    {
-        urgencyLevel: 2,
-        itemName: "Nitrile Gloves",
-        quantity: "1,000",
-        budget: "Need Donate",
-        specifications: "New Orleans",
-        createdAt: "Danile"
-    },
-    {
-        urgencyLevel: 3,
-        itemName: "Face shield",
-        quantity: "500",
-        budget: "Can buy",
-        specifications: "Paris",
-        createdAt: "Nursing Home"
-    },
-    {
-        urgencyLevel: 3,
-        itemName: "Face shield",
-        quantity: "500",
-        budget: "Can buy",
-        specifications: "Paris",
-        createdAt: "Nursing Home"
-    },
-    {
-        urgencyLevel: 3,
-        itemName: "Ventilators",
-        quantity: "20",
-        budget: "Can buy",
-        specifications: "Madrid",
-        createdAt: "Health Center"
-    },
-    {
-        urgencyLevel: 4,
-        itemName: "Nitrile Gloves",
-        quantity: "1,000",
-        budget: "Need Donate",
-        specifications: "Madrid",
-        createdAt: "Health Center"
-    }
-]
-
-export const leadsDummy = [
-    {
-        id : 1,
-        itemName: "Surgical Masks",
-        specifications: "Manufacturer",
-        quantity: "500,000",
-        cost: "Sell",
-        createdAt: "Hong Kong",
-        availableAt: "Factory Ltd",
-    },
-    {
-        id: 2,
-        itemName: "Face Shields",
-        specifications: "NGO",
-        quantity: "3,000",
-        cost: "Donate",
-        createdAt: "Ning bo",
-        availableAt: "Medical Aid Assol",
-    },
-    {
-        id: 3,
-        itemName: "3-Ply Masks",
-        specifications: "Distributor",
-        quantity: "750,000",
-        cost: "Sell",
-        createdAt: "Kyoto",
-        availableAt: "Agent Co.Ltd",
-    },
-    {
-        id: 4,
-        itemName: "N95 Masks",
-        specifications: "Distributor",
-        quantity: "150",
-        cost: "Donate",
-        createdAt: "Omaka",
-        availableAt: "Grorge",
-    },
-    {
-        id: 5,
-        itemName: "Surgical Masks",
-        specifications: "Distributor",
-        quantity: "150,00",
-        cost: "Donate",
-        createdAt: "Hong Kong",
-        availableAt: "Factory Ltd",
-    }
-]
